Making MS compiler happy

+++ crypto_kem/kyber512/clean/symmetric.h
@@ -9,8 +9,8 @@ typedef struct {
 } keccak_state;
 
 void PQCLEAN_KYBER512_CLEAN_kyber_shake128_absorb(keccak_state *s, const unsigned char *input, unsigned char x, unsigned char y);
-void PQCLEAN_KYBER512_CLEAN_kyber_shake128_squeezeblocks(unsigned char *output, unsigned long long nblocks, keccak_state *s);
-void PQCLEAN_KYBER512_CLEAN_shake256_prf(unsigned char *output, unsigned long long outlen, const unsigned char *key, unsigned char nonce);
+void PQCLEAN_KYBER512_CLEAN_kyber_shake128_squeezeblocks(unsigned char *output, size_t nblocks, keccak_state *s);
+void PQCLEAN_KYBER512_CLEAN_shake256_prf(unsigned char *output, size_t outlen, const unsigned char *key, unsigned char nonce);
 
 #define hash_h(OUT, IN, INBYTES) sha3_256(OUT, IN, INBYTES)
 #define hash_g(OUT, IN, INBYTES) sha3_512(OUT, IN, INBYTES)
